# Improve native_batch_norm_backward performance (CUDA)