--- a/aten/src/ATen/native/cuda/Normalization.cu
+++ b/aten/src/ATen/native/cuda/Normalization.cu
@@ -11,7 +11,8 @@ namespace at { namespace native {
 namespace {
 
 inline bool batch_norm_use_channels_last_kernels(const at::Tensor& self) {
-  return self.is_contiguous(at::MemoryFormat::ChannelsLast) || self.ndimension() == 2;
+  return (self.is_contiguous(at::MemoryFormat::ChannelsLast) ||
+          (self.is_contiguous() && self.strides()[1] == 1));
 }
 
 enum class Impl {
@@ -34,6 +35,15 @@ inline Impl batch_norm_choose_impl(const Tensor& self) {
   }
 
   return Impl::General;
+}
+
+inline Impl batch_norm_choose_impl(const Tensor& in1, const Tensor& in2) {
+  auto imp1 = batch_norm_choose_impl(in1);
+  if (imp1 == Impl::General) {
+    return imp1;
+  }
+  auto imp2 = batch_norm_choose_impl(in2);
+  return imp1 == imp2 ? imp1 : Impl::General;
 }
 
 void batch_norm_elementwise(
@@ -105,6 +115,125 @@ void batch_norm_elementwise(
   }
   }
 }
+
+Tensor batch_norm_elementwise_backward_train(
+    const Tensor& grad_out, const Tensor& input, const Tensor& mean, const Tensor& invstd,
+    const Tensor& weight, const Tensor& sum_dy, const Tensor& sum_dy_xmu) {
+  switch (batch_norm_choose_impl(input, grad_out)) {
+  case Impl::Contiguous: {
+    return AT_DISPATCH_FLOATING_TYPES_AND2(kHalf, kBFloat16, input.scalar_type(),
+                                           "batch_norm_backward_elemt", [&] {
+      return batch_norm_backward_elemt_cuda_template<scalar_t, scalar_t, int32_t>(
+          grad_out, input, mean, invstd, weight, sum_dy, sum_dy_xmu);
+    });
+  }
+  case Impl::ChannelsLast: {
+    if ((!weight.defined() || weight.is_contiguous()) &&
+        mean.is_contiguous() && invstd.is_contiguous()) {
+      return batch_norm_backward_elemt_channels_last_cuda_template(
+          grad_out, input, mean, invstd, weight, sum_dy, sum_dy_xmu);
+    }
+    C10_FALLTHROUGH;
+  }
+  case Impl::General: {
+    const auto ndim = input.dim();
+    DimVector sizes(ndim, 1), strides(ndim, 0);
+    auto as_nd = [&](const Tensor& t) {
+      TORCH_INTERNAL_ASSERT(t.defined() && t.dim() == 1);
+      sizes[1] = t.sizes()[0];
+      strides[1] = t.strides()[0];
+      return t.as_strided(sizes, strides);
+    };
+    auto invstd_nd = as_nd(invstd);
+    auto mean_nd = as_nd(mean);
+    auto sum_dy_nd = as_nd(sum_dy);
+    auto sum_dy_xmu_nd = as_nd(sum_dy_xmu);
+    auto weight_nd = weight.defined() ? as_nd(weight) :
+        at::scalar_tensor(1.0, input.options().dtype(mean.scalar_type()));
+
+    Tensor grad_input = at::empty(input.sizes(), grad_out.options());
+    auto iter = TensorIteratorConfig()
+        .add_output(grad_input)
+        .add_input(grad_out)
+        .add_input(input)
+        .add_input(weight_nd)
+        .add_input(mean_nd)
+        .add_input(invstd_nd)
+        .add_input(sum_dy_xmu_nd)
+        .add_input(sum_dy_nd)
+        .check_all_same_dtype(false)
+        .promote_inputs_to_common_dtype(false)
+        .build();
+
+    AT_DISPATCH_FLOATING_TYPES_AND2(kHalf, kBFloat16, grad_out.scalar_type(),
+                                    "batch_norm_eval_backward", [&]{
+      using accscalar_t = at::acc_type<scalar_t, true>;
+      auto norm_fct = static_cast<accscalar_t>(1.0 / (input.numel() /input.size(1)) );
+      gpu_kernel(iter, [norm_fct] GPU_LAMBDA (scalar_t gO, scalar_t input, accscalar_t weight,
+                                              accscalar_t mean, accscalar_t invstd,
+                                              accscalar_t xmu, accscalar_t dy) -> scalar_t {
+        auto factor_1_c = invstd * invstd * xmu * norm_fct;
+        auto factor_2_c = weight * invstd;
+        auto m_dy_c = dy * norm_fct;
+        return (gO - m_dy_c - (input - mean) * factor_1_c) * factor_2_c;
+      });
+    });
+    return grad_input;
+  }
+  }
+  TORCH_INTERNAL_ASSERT(false);
+}
+
+Tensor batch_norm_elementwise_backward_eval(
+    const Tensor& grad_out, const Tensor& input,
+    const Tensor& invstd, const Tensor& weight) {
+  const auto ndim = input.dim();
+  DimVector shape(ndim, 1), strides(ndim, 0);
+  shape[1] = invstd.sizes()[0];
+  strides[1] = invstd.strides()[0];
+  auto invstd_nd = invstd.as_strided(shape, strides);
+  Tensor grad_input = at::empty(input.sizes(), grad_out.options());
+
+  if (weight.defined()) {
+    strides[1] = weight.strides()[0];
+    auto weight_nd = weight.as_strided(shape, strides);
+    auto iter = TensorIteratorConfig()
+        .add_output(grad_input)
+        .add_input(grad_out)
+        .add_input(invstd_nd)
+        .add_input(weight_nd)
+        .check_all_same_dtype(false)
+        .promote_inputs_to_common_dtype(false)
+        .build();
+
+    AT_DISPATCH_FLOATING_TYPES_AND2(kHalf, kBFloat16, grad_out.scalar_type(),
+                                    "batch_norm_eval_backward", [&]{
+      using accscalar_t = at::acc_type<scalar_t, true>;
+      gpu_kernel(iter, [] GPU_LAMBDA (scalar_t gO, accscalar_t invstd, accscalar_t weight)
+                 -> scalar_t {
+          return gO * weight * invstd;
+      });
+    });
+  } else {
+    auto iter = TensorIteratorConfig()
+        .add_output(grad_input)
+        .add_input(grad_out)
+        .add_input(invstd_nd)
+        .check_all_same_dtype(false)
+        .promote_inputs_to_common_dtype(false)
+        .build();
+
+    AT_DISPATCH_FLOATING_TYPES_AND2(kHalf, kBFloat16, grad_out.scalar_type(),
+                                    "batch_norm_eval_backward", [&]{
+      using accscalar_t = at::acc_type<scalar_t, true>;
+      gpu_kernel(iter, [] GPU_LAMBDA (scalar_t gO, accscalar_t invstd) -> scalar_t {
+          return gO * invstd;
+      });
+    });
+  }
+  return grad_input;
+}
+
 
 void batch_norm_mean_var(const Tensor& self, Tensor& save_mean, Tensor& save_var) {
   // NOTE: Epsilon is only used for InvStd, not Var. The value here is ignored.
@@ -284,35 +413,72 @@ std::tuple<Tensor, Tensor, Tensor> batch_norm_cuda(const Tensor& self, const c10
   return std::make_tuple(output, save_mean, save_invstd);
 }
 
-std::tuple<Tensor, Tensor, Tensor> batch_norm_backward_cuda(const Tensor& grad_out, const Tensor& self, const c10::optional<Tensor>& weight_opt, const c10::optional<Tensor>& running_mean_opt, const c10::optional<Tensor>& running_var_opt, const c10::optional<Tensor>& save_mean_opt, const c10::optional<Tensor>& save_invstd_opt, bool train, double epsilon, std::array<bool,3> grad_input_mask) {
+std::tuple<Tensor, Tensor, Tensor> batch_norm_backward_cuda(const Tensor& grad_out, const Tensor& input, const c10::optional<Tensor>& weight_opt, const c10::optional<Tensor>& running_mean_opt, const c10::optional<Tensor>& running_var_opt, const c10::optional<Tensor>& save_mean_opt, const c10::optional<Tensor>& save_invstd_opt, bool train, double epsilon, std::array<bool,3> grad_input_mask) {
   // See [Note: hacky wrapper removal for optional tensor]
-  c10::MaybeOwned<Tensor> weight_maybe_owned = at::borrow_from_optional_tensor(weight_opt);
-  const Tensor& weight = *weight_maybe_owned;
-  const Tensor& running_mean = c10::value_or_else(running_mean_opt, [] {return Tensor();});
-  const Tensor& running_var = c10::value_or_else(running_var_opt, [] {return Tensor();});
-  const Tensor& save_mean = c10::value_or_else(save_mean_opt, [] {return Tensor();});
-  const Tensor& save_invstd = c10::value_or_else(save_invstd_opt, [] {return Tensor();});
+  c10::MaybeOwned<Tensor> weight = at::borrow_from_optional_tensor(weight_opt);
+  c10::MaybeOwned<Tensor> save_mean = at::borrow_from_optional_tensor(save_mean_opt);
+  c10::MaybeOwned<Tensor> save_invstd = at::borrow_from_optional_tensor(save_invstd_opt);
+  c10::MaybeOwned<Tensor> running_mean = at::borrow_from_optional_tensor(running_mean_opt);
+  c10::MaybeOwned<Tensor> running_var = at::borrow_from_optional_tensor(running_var_opt);
 
-  return AT_DISPATCH_FLOATING_TYPES_AND2(at::ScalarType::Half, at::ScalarType::BFloat16, self.scalar_type(), "batch_norm_backward_cuda", [&] {
-    auto mean_st = running_mean.dtype();
-    auto var_st = running_var.dtype();
-    TORCH_CHECK(mean_st == var_st, "running_mean and running_var need to have the same data types");
-    bool is_half_float = std::is_same<scalar_t, at::Half>::value && mean_st == at::kFloat;
-    bool is_bfloat16_float = std::is_same<scalar_t, at::BFloat16>::value && mean_st == at::kFloat;
-    if (cuda::detail::canUse32BitIndexMath(self)) {
-      if (is_half_float || is_bfloat16_float) {
-        return batch_norm_backward_cuda_template<scalar_t, float, int32_t>(grad_out, self, weight, running_mean, running_var, save_mean, save_invstd, train, epsilon, grad_input_mask);
-      } else {
-        return batch_norm_backward_cuda_template<scalar_t, scalar_t, int32_t>(grad_out, self, weight, running_mean, running_var, save_mean, save_invstd, train, epsilon, grad_input_mask);
-      }
+  const bool needs_reduction = train || grad_input_mask[1] || grad_input_mask[2];
+
+  // Fused reducion & elementwise kernel
+  if (needs_reduction && grad_input_mask[0] &&
+      !batch_norm_use_channels_last_kernels(input) &&
+      cuda::detail::canUse32BitIndexMath(input) &&
+      cuda::detail::canUse32BitIndexMath(grad_out)) {
+    return AT_DISPATCH_FLOATING_TYPES_AND2(kHalf, kBFloat16, input.scalar_type(),
+                                           "batch_norm_backward_cuda", [&] {
+      return batch_norm_backward_cuda_template<scalar_t, scalar_t, int32_t>(
+          grad_out, input, *weight, *running_mean, *running_var,
+          *save_mean, *save_invstd, train, epsilon, grad_input_mask);
+    });
+  }
+
+  // NOTE: native_batch_norm always returns save_mean and save_invstd to be reused in backward.
+  // However, this is also called from cudnn_batch_norm in eval mode which doesn't give
+  // save_mean and save_invstd, so it needs recalculated.
+  const auto acc_type = at::toAccumulateType(input.scalar_type(), /*is_cuda=*/true);
+  Tensor mean;
+  if (save_mean->defined()) {
+    mean = *save_mean;
+  } else if (needs_reduction) {
+    TORCH_CHECK(running_mean->defined());
+    mean = (running_mean->scalar_type() == acc_type) ?
+        *running_mean : running_mean->to(acc_type);
+  }
+
+  Tensor invstd;
+  if (save_invstd->defined()) {
+    invstd = *save_invstd;
+  } else {
+    TORCH_CHECK(running_var->defined());
+    auto n_channels = input.sizes()[1];
+    invstd = at::empty({n_channels}, input.options().dtype(acc_type));
+    batch_norm_calc_invstd(invstd, *running_var, epsilon);
+  }
+
+  Tensor sum_dy, sum_dy_xmu, grad_weight, grad_bias;
+  if (needs_reduction) {
+    std::tie(sum_dy, sum_dy_xmu, grad_weight, grad_bias) =
+        batch_norm_backward_reduce_cuda(
+            grad_out, input, mean, invstd, *weight,
+            grad_input_mask[0], grad_input_mask[1], grad_input_mask[2]);
+  }
+
+  Tensor grad_input;
+  if (grad_input_mask[0]) {
+    if (train) {
+      grad_input = batch_norm_elementwise_backward_train(
+          grad_out, input, mean, invstd, *weight, sum_dy, sum_dy_xmu);
     } else {
-      if (is_half_float || is_bfloat16_float) {
-        return batch_norm_backward_cuda_template<scalar_t, float, int64_t>(grad_out, self, weight, running_mean, running_var, save_mean, save_invstd, train, epsilon, grad_input_mask);
-      } else {
-        return batch_norm_backward_cuda_template<scalar_t, scalar_t, int64_t>(grad_out, self, weight, running_mean, running_var, save_mean, save_invstd, train, epsilon, grad_input_mask);
-      }
+      grad_input = batch_norm_elementwise_backward_eval(
+          grad_out, input, invstd, *weight);
     }
-  });
+  }
+
+  return std::make_tuple(grad_input, grad_weight, grad_bias);
 }
 
 std::tuple<Tensor, Tensor> batch_norm_stats_cuda(const Tensor& self, double epsilon) {
@@ -391,33 +557,39 @@ std::tuple<Tensor, Tensor> batch_norm_gather_stats_with_counts_cuda(
   });
 }
 
-std::tuple<Tensor, Tensor, Tensor, Tensor> batch_norm_backward_reduce_cuda(const Tensor& self, const Tensor& input, const Tensor& mean, const Tensor& invstd, const c10::optional<Tensor>& weight_opt, bool input_g, bool weight_g, bool bias_g) {
+std::tuple<Tensor, Tensor, Tensor, Tensor> batch_norm_backward_reduce_cuda(const Tensor& grad_output, const Tensor& input, const Tensor& mean, const Tensor& invstd, const c10::optional<Tensor>& weight_opt, bool input_g, bool weight_g, bool bias_g) {
   // See [Note: hacky wrapper removal for optional tensor]
   c10::MaybeOwned<Tensor> weight_maybe_owned = at::borrow_from_optional_tensor(weight_opt);
   const Tensor& weight = *weight_maybe_owned;
 
-  // self is grad_output
-  if (at::cuda::detail::canUse32BitIndexMath(self) && batch_norm_use_channels_last_kernels(self)){
-    return batch_norm_backward_reduce_cuda_channels_last_template(self, input, mean, invstd, weight, input_g, weight_g, bias_g);
+  if (at::cuda::detail::canUse32BitIndexMath(grad_output) &&
+      batch_norm_use_channels_last_kernels(grad_output) &&
+      batch_norm_use_channels_last_kernels(input) &&
+      (!weight.defined() || weight.is_contiguous()) &&
+      mean.is_contiguous() && invstd.is_contiguous()){
+    return batch_norm_backward_reduce_cuda_channels_last_template(
+        grad_output, input, mean, invstd, weight, input_g, weight_g, bias_g);
   }
 
-  return AT_DISPATCH_FLOATING_TYPES_AND2(at::ScalarType::Half, at::ScalarType::BFloat16, self.scalar_type(), "batch_norm_backward_reduce", [&] {
+  return AT_DISPATCH_FLOATING_TYPES_AND2(at::ScalarType::Half, at::ScalarType::BFloat16, grad_output.scalar_type(), "batch_norm_backward_reduce", [&] {
     auto mean_st = mean.dtype();
     auto invstd_st = invstd.dtype();
     TORCH_CHECK(mean_st == invstd_st, "mean and invstd need to have the same data types");
-    bool is_half_float = std::is_same<scalar_t, at::Half>::value && mean_st == at::kFloat;
-    bool is_bfloat16_float = std::is_same<scalar_t, at::BFloat16>::value && mean_st == at::kFloat;
-    if (cuda::detail::canUse32BitIndexMath(self)) {
-      if (is_half_float || is_bfloat16_float) {
-        return batch_norm_backward_reduce_cuda_template<scalar_t, float, int32_t>(self, input, mean, invstd, weight, input_g, weight_g, bias_g);
+    const bool is_half_float = weight.defined() &&
+        weight.scalar_type() != input.scalar_type();
+    using accscalar_t = at::acc_type<scalar_t, true>;
+
+    if (cuda::detail::canUse32BitIndexMath(grad_output)) {
+      if (is_half_float) {
+        return batch_norm_backward_reduce_cuda_template<scalar_t, accscalar_t, int32_t>(grad_output, input, mean, invstd, weight, input_g, weight_g, bias_g);
       } else {
-        return batch_norm_backward_reduce_cuda_template<scalar_t, scalar_t, int32_t>(self, input, mean, invstd, weight, input_g, weight_g, bias_g);
+        return batch_norm_backward_reduce_cuda_template<scalar_t, scalar_t, int32_t>(grad_output, input, mean, invstd, weight, input_g, weight_g, bias_g);
       }
     } else {
-      if (is_half_float || is_bfloat16_float) {
-        return batch_norm_backward_reduce_cuda_template<scalar_t, float, int64_t>(self, input, mean, invstd, weight, input_g, weight_g, bias_g);
+      if (is_half_float) {
+        return batch_norm_backward_reduce_cuda_template<scalar_t, accscalar_t, int64_t>(grad_output, input, mean, invstd, weight, input_g, weight_g, bias_g);
       } else {
-        return batch_norm_backward_reduce_cuda_template<scalar_t, scalar_t, int64_t>(self, input, mean, invstd, weight, input_g, weight_g, bias_g);
+        return batch_norm_backward_reduce_cuda_template<scalar_t, scalar_t, int64_t>(grad_output, input, mean, invstd, weight, input_g, weight_g, bias_g);
       }
     }
   });