@ngimel 2021-05-13
Awesome! Can you add benchmarks similar to original issue, very large N and small C, (125000, 5000) or something like that?