@peterbell10 2021-05-13
Sure, here are the shapes from the issue
|shape|old (us)|new (us)|
|-|-|-|
|(125000, 50)|6440|639|
|(50, 50, 50, 50)|1100|1090|